refactor: move token verification logic from tokenMiddlewares to userMiddlewares
